Fix: изменение layout
