+++ src/components/AdCard/AdCard.tsx
@@ -3,7 +3,7 @@ import { Text } from '@consta/uikit/Text';
 
 const AdCard = () => {
     return (
-      <Card verticalSpace="5xl" horizontalSpace="5xl" style={{height: 190, width: "165vh", background:" #777777", borderRadius: "0px 0px 16px 16px"}}>
+      <Card verticalSpace="5xl" horizontalSpace="5xl" style={{height: 190, width: "84vw", background:" #777777", borderRadius: "0px 0px 16px 16px"}}>
         <Text>Тут будет реклама (возможно онлайн казино)</Text>
       </Card>
     );
