+++ src/components/Header/Header.scss
@@ -1,0 +1,7 @@
+.Search .TextField-InputContainer_view_default{
+    width: 310px;
+    height: 20px;
+    margin: 8;
+    background: #FFFBF5;
+    border: none;
+}
