+++ src/components/NavMenu/Menu.scss
@@ -3,6 +3,8 @@
     width: 237px;
     height: max;
     gap: 16px;
+    min-height: 100vh;
+    position: fixed;
 }
 
 .text-on-top{
