+++ src/components/pages/MainPage/MainPage.scss
@@ -1,11 +1,3 @@
-.Search .TextField-InputContainer_view_default{
-    width: 310px;
-    height: 20px;
-    margin: 8;
-    background: #FFFBF5;
-    border: none;
-}
-
 .division-text{
     font-family: Inter;
     font-size: 19px;
