+++ src/components/pages/NotFoundPage/NotFoundPage.tsx
@@ -1,5 +1,12 @@
 const NotFound = () => {
-    return <h1>Not Found</h1>
+    return (
+      <div style={{height: "80vh", padding: 20}}>
+        <h1>
+          Not Found
+        </h1>
+      </div>
+      
+      )
   }
   
   export default NotFound
